+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ New features
 * Add parameter scatter matrix plot to bootstrap
 * Add :code:`modeling.insert_ebes_into_dataset`
 * Add :code:`modeling.add_placebo_model`
+* Add :code:`modeling.binarize_dataset`
 
 Changes
 =======